fix: P2-01,P2-02,P3-01,P3-03,P3-04: quality costs int, bare excepts, schema, tests, CI
- P2-01: QUALITY_COSTS float→int (2.5→2, 3.5→4), DRY: generation.py imports from quality_pricing
- P2-02: 3x bare except→except Exception (generation.py:5750, main.py:3389, common.py:4868)
- P3-01: schema_postgres.sql credits REAL→INTEGER (users, feed, remix tables)
- P3-03: .github/workflows/ci.yml with lint+test+live stages
- P3-04: characterization tests for quality costs + no-bare-except test
- New: test_no_bare_excepts.py (AST-based, finds any bare except in bot/)

+++ schema_postgres.sql
@@ -10,7 +10,7 @@ BEGIN;
 CREATE TABLE IF NOT EXISTS users (
     id BIGSERIAL PRIMARY KEY,
     telegram_id BIGINT UNIQUE NOT NULL,
-    credits REAL DEFAULT 0,
+    credits INTEGER DEFAULT 0,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     referral_code TEXT,
@@ -294,8 +294,8 @@ CREATE TABLE IF NOT EXISTS feed_remix_events (
     remix_generation_task_id BIGINT NOT NULL,
     source_author_id BIGINT NOT NULL,
     remix_author_id BIGINT NOT NULL,
-    credits_spent REAL DEFAULT 0,
-    royalty_credits REAL DEFAULT 0,
+    credits_spent INTEGER DEFAULT 0,
+    royalty_credits INTEGER DEFAULT 0,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     UNIQUE(source_generation_task_id, remix_generation_task_id)
 );
@@ -310,7 +310,7 @@ CREATE TABLE IF NOT EXISTS prompt_repeat_events (
     source_type TEXT NOT NULL,
     source_id BIGINT NOT NULL,
     repeat_task_id TEXT,
-    credits_spent REAL DEFAULT 0,
+    credits_spent INTEGER DEFAULT 0,
     amount_rub REAL NOT NULL DEFAULT 10,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
